GL-14: Add logic for player login and more
Add methods to login, logout and create an account for the self player
as well as dynamic fetching of account details from the API in these
cases. Storage and access of the API JWT token fine tuned. Support for
anonymous self player (playing without an account) is improved thanks to
optimizations during API detail fetching and saving them to local file.

Add accessors in JoueurManager to access inner structures of the manager
class. These are a work in progress and don't currently abide by the
encapsulation rules (the returned references are not clones).

Code organization, cleanup and more comments to illustrate the current
issues, todos and shortcomings of the implementation.